backend/requirementsaa.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py==1.0.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
# backports==1.0
backports.functools_lru_cache==1.6.4
beautifulsoup4==4.11.1
bleach==5.0.0
# bzip2==1.0.8
# c-ares==1.18.1
# ca-certificates==2022.5.18.1
cached-property==1.5.2
cached_property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
flask==2.1.2
flask-cors==3.0.10
flatbuffers==2.0
flit-core==3.7.1
fonttools==4.31.2
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.45.0
h5py==3.6.0
# hdf5==1.12.1
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
importlib_resources==5.7.1
ipykernel==6.10.0
ipython==8.2.0
ipython_genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
jinja2==3.1.2
jsonschema==4.5.1
jupyter==1.0.0
jupyter_client==7.2.0
jupyter_console==6.4.3
jupyter_core==4.9.2
jupyterlab_pygments==0.2.2
jupyterlab_widgets==1.1.0
keras==2.8.0
keras-preprocessing==1.1.2
kiwisolver==1.4.2
# krb5==1.19.3
# libblas==3.9.0
# libcblas==3.9.0
libclang==13.0.0
libcurl==7.82.0
libcxx==13.0.1
libedit==3.1.20191231
libev==4.33
libffi==3.4.2
libgfortran==5.0.0.dev0
libgfortran5==11.0.1.dev0
liblapack==3.9.0
libnghttp2==1.47.0
libopenblas==0.3.18
libsodium==1.0.18
libssh2==1.10.0
libzlib==1.2.11
llvm-openmp==13.0.1
markdown==3.3.6
markupsafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.4.0
ncurses==6.3
nest-asyncio==1.5.4
networkx==2.7.1
notebook==6.4.11
numpy==1.21.5
oauthlib==3.2.0
openssl==3.0.3
opt-einsum==3.3.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.0.1
pip==22.0.4
prometheus_client==0.14.1
prompt-toolkit==3.0.29
prompt_toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python==3.8.13
python-dateutil==2.8.2
python-fastjsonschema==2.15.3
python_abi==3.8==
pywavelets==1.3.0==
pyzmq==22.3.0
readline==8.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
send2trash==1.8.0
setuptools==61.2.0
six==1.15.0
soupsieve==2.3.1
sqlite==3.37.1
stack_data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-deps==2.8.0==0
tensorflow-macos==2.8.0
tensorflow-metal==0.4.0
termcolor==1.1.0
terminado==0.15.0
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2022.3.25
tk==8.6.12
tornado==6.1
traitlets==5.1.1
typing-extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==2.1.0
wheel==0.37.1
widgetsnbextension==3.6.0
wrapt==1.14.0
xz==5.2.5
zeromq==4.3.4
zipp==3.7.0
zlib==1.2.11
